mturksegutils/requirements.txt
boto3 == 1.28.16
botocore == 1.31.16
Pillow == 9.0.1
numpy == 1.21.5
xmltodict == 0.13.0
flask == 2.2.3
flask_socketio == 5.3.6
selenium == 4.11.2
matplotlib == 3.5.2
pandas == 0.24.2